Make popup filter show matching fields always
Previously, if a field of a NS record happened to be an object (e.g. an
Array) and none of its elements matched the search string (typed by the
user in the popup window), then this field would not appear in the
filtered result, misleading the user to think that the field was not
present.

Now the field will always be added to the filtered popup contents, if
it matches the search string